requirements.txt
cython==0.29.20
pandas==1.0.5
certifi==2020.6.20
cryptography
chardet==3.0.4
idna==2.10
azure-core
azure-common
azure-nspkg
azure-functions
azure-storage-common
azure-storage-nspkg
azure-storage-blob
azure-storage-queue
azure-servicebus
azure-cosmos
azure-mgmt-servicebus
azure-functions-durable
redis==3.5.3
PyYAML==5.4.1
jinja2==2.11.3
requests==2.25.1
dicttoxml==1.7.4
docxtpl==0.11.3
reportlab==3.5.59
plotly==4.14.3
kaleido==0.1.0
svglib==1.0.1
lxml==4.6.2
tables==3.6.1
pyarrow==3.0.0
sqlalchemy==1.3.23
psycopg2-binary
orjson
asyncpg
opencensus-ext-azure
opencensus-ext-logging
opencensus-ext-requests
matplotlib
